+++ common/src/main/java/com/yanny/ali/plugin/common/nodes/DynamicNode.java
@@ -26,7 +26,7 @@ public class DynamicNode implements IDataNode {
         List<LootItemFunction> allFunctions = Stream.concat(functions.stream(), Arrays.stream(entry.functions)).toList();
         List<LootItemCondition> allConditions = Stream.concat(conditions.stream(), Arrays.stream(entry.conditions)).toList();
 
-        tooltip = EntryTooltipUtils.getDynamicTooltip(entry, chance, sumWeight);
+        tooltip = EntryTooltipUtils.getDynamicTooltip(utils, entry, chance, sumWeight, allFunctions, allConditions);
     }
 
     public DynamicNode(IClientUtils utils, FriendlyByteBuf buf) {